adds methods for deleting item and getting a single item also adds item activity and pie chart

+++ app/src/main/java/com/prateekthakur272/bunkmate/ItemActivity.kt
@@ -1,9 +1,12 @@
 package com.prateekthakur272.bunkmate
 
+import android.content.DialogInterface
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
+import androidx.appcompat.app.AlertDialog
+import com.google.android.material.snackbar.Snackbar
 import com.prateekthakur272.bunkmate.databinding.ActivityItemBinding
 import org.eazegraph.lib.models.PieModel
 import kotlin.math.roundToInt
@@ -21,17 +24,16 @@ class ItemActivity : AppCompatActivity() {
         itemDatabaseHelper = ItemDatabaseHelper(this)
         subject = itemDatabaseHelper.getItem(intent.getIntExtra("id",-1))
         supportActionBar?.title = subject.title
-
-        binding.pieChart.addPieSlice(
-            PieModel(
-                subject.lectureAttended.toFloat(),
-                this.getColor(R.color.green)
-            )
-        )
         binding.pieChart.addPieSlice(
             PieModel(
                 (subject.totalLectures-subject.lectureAttended).toFloat(),
                 this.getColor(R.color.red)
+            )
+        )
+        binding.pieChart.addPieSlice(
+            PieModel(
+                subject.lectureAttended.toFloat(),
+                this.getColor(R.color.green)
             )
         )
 
@@ -45,12 +47,18 @@ class ItemActivity : AppCompatActivity() {
         menuInflater.inflate(R.menu.item_menu,menu)
         return true
     }
-
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when(item.itemId){
             R.id.delete ->{
-                itemDatabaseHelper.deleteItem(subject.id)
-                finish()
+                val dialog = AlertDialog.Builder(this)
+                dialog.setTitle("Do you want to delete ${subject.title}?")
+                dialog.setCancelable(false)
+                dialog.setPositiveButton("Delete", DialogInterface.OnClickListener { _, i ->
+                    itemDatabaseHelper.deleteItem(subject.id)
+                    finish()
+                })
+                dialog.setNegativeButton("No", DialogInterface.OnClickListener { _, _ ->})
+                dialog.create().show()
                 return true
             }
         }
